Absorb flag package into root
This is the first step in collapsing `clip` into being a single package.
I don't think asking users to keep track of three names, two of which
squat on reasonable words, is better than housing all the logic under a
single package.

The README is documented with what's effectively the full plan at this
point, although some names will be tweaked a bit I'm sure. The command
package is next.