Use parent writer when possible
The idea is that if a sub-command has its own writer specified, use
that. If not, use any specified parent writer before using the default.

I suspect that folks will only realistically ever declare this at the
top-level, but if the option is available at every level, might as well
cut out the foot guns.