Simplify command nesting
This follows the lead of flag creation, where the return value of the
constructor of a command (`clip.NewCommand`) is meant to be called
inline as the sole argument to another constructor (`clip.SubCommand`).
If that's the case, why not inline the args?

I suppose the counter-argument is that it might be useful to have a
direct reference to each sub-command. If this proves to be true, I
wouldn't be surprised if this ends up getting undone, especially when
we're considering something like how to check if flags have been set.
But for now, I'm going to keep it simple and keep it consistent.